src/test/java/learn/springframework/controllers/SetterInjectedControllerTest.java
package learn.springframework.controllers;

import static org.junit.Assert.assertEquals;

import org.junit.Before;
import org.junit.Test;

import learn.springframework.services.GreetingServiceImpl;

public class SetterInjectedControllerTest {
	private SetterInjectedController setterInjectedController;

	@Before
	public void setUp() throws Exception {
		this.setterInjectedController = new SetterInjectedController();
		this.setterInjectedController.setGreetingService(new GreetingServiceImpl());
	}

	@Test
	public void testSayHello() {
		assertEquals(GreetingServiceImpl.SAY_HELLO, setterInjectedController.sayHello());
	}

}
